remove unnecessary templates
