+++ performance_test_factory/examples/client_nodes_main.cpp
@@ -99,7 +99,7 @@ int main(int argc, char ** argv)
     rclcpp::init(argc, argv);
 
     performance_test::TemplateFactory<> factory(use_ipc, use_ros_params, verbose, ros_namespace);
-    performance_test::System<> ros2_system(static_cast<performance_test::ExecutorType>(executor));
+    performance_test::System ros2_system(static_cast<performance_test::ExecutorType>(executor));
     ros2_system.enable_events_logger(events_file_path);
 
     std::vector<std::shared_ptr<performance_test::PerformanceNode<>>> client_nodes =
