+++ performance_test_factory/examples/json_system_main.cpp
@@ -84,7 +84,7 @@ int main(int argc, char** argv)
     rclcpp::init(argc, argv);
 
     // Architecture
-    performance_test::System<> ros2_system(static_cast<performance_test::ExecutorType>(executor));
+    performance_test::System ros2_system(static_cast<performance_test::ExecutorType>(executor));
     ros2_system.enable_events_logger(events_output_path);
 
     performance_test::TemplateFactory<> factory(use_ipc);
